requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-extensions==3.2.3
django-debug-toolbar==4.4.6
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.15.2
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
et_xmlfile==2.0.0
idna==3.10
mssql-django==1.4
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
